DistributedLogger (Distributed Logger v0.1.0)

Functions

generate_consolidated_file(initial_line, final_line)

Specs

generate_consolidated_file(integer(), integer()) ::
  {:ok, String.t()} | {:error, String.t(), [atom()]}

Generated a consilidate event log file, which guarantee consistency along all nodes

Partial retrieves (initial line > 0 or final line < last line) may be inconsistent, but can be used for performance reasons

border inclusive

is safe to use any integer as parameter

To retrieve all file you just need to use 0 as initial line and a very big number as final line

The file is generated with a time stamp and inside the node's data folder

The file name is returned as second element of the tuple

Examples

iex> {:ok, _file_path} = DistributedLogger.generate_consolidated_file(0,100)

read_local(initial_line, final_line)

Specs

read_local(integer(), integer()) :: [String.t()]

Read lines of the local log file.

0 index based

border inclusive

is safe to use any integer as parameter

Examples

iex> DistributedLogger.write_local("event 0")
iex> DistributedLogger.write_local("event 1")
iex> DistributedLogger.write_local("event 2")
iex> DistributedLogger.write_local("event 3")
iex> DistributedLogger.read_local(1,2)
iex> ["event 1", "event 2"]

iex> DistributedLogger.read_local(1,10)
iex> []

iex> DistributedLogger.write_local("event 0")
iex> DistributedLogger.write_local("event 1")
iex> DistributedLogger.read_local(-57,454785)
iex> ["event 0", "event 1"]

write_global(event_data)

Specs

write_global(String.t()) :: :ok

Writes event data globally (all nodes)

Examples

iex> DistributedLogger.write_global("event data")
iex> :ok

write_local(event_data)

Specs

write_local(String.t()) :: :ok

Writes event data locally (only this node)

Examples

iex> DistributedLogger.write_local("event data")
iex> :ok